pricetrack/requirements.txt
aiohttp
asyncio

# wsgi
gunicorn

# CORS
aiohttp_cors

# data validation
cerberus

# mongodb driver
motor

# JWT token authorization
PyJWT

# password hashing
bcrypt

# async tasks
celery
requests
itsdangerous

# bs4 for data parsing
beautifulsoup4

# testing
pytest
pytest-aiohttp
